% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_rot.R
\name{bandwidth_rot}
\alias{bandwidth_rot}
\title{MASS' Rule of Thumb Bandwidth Estimation}
\usage{
bandwidth_rot(x)
}
\arguments{
\item{x}{Type: numeric vector. A data vector.}
}
\value{
A bandwidth on a scale suitable for the \code{width} argument of \code{density}.
}
\description{
A well-supported rule-of-thumb for choosing the bandwidth of a Gaussian kernel density estimator.
}
\examples{
\dontrun{
#?MASS::bandwidth.nrd
bandwidth_rot <- function (x) {
  r <- quantile(x, c(0.25, 0.75))
  h <- (r[2L] - r[1L])/1.34
  4 * 1.06 * min(sqrt(var(x)), h) * length(x)^(-1/5)
}
}

}

