% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindlist.R
\name{cbindlist}
\alias{cbindlist}
\title{data.table column list binding}
\usage{
cbindlist(target)
}
\arguments{
\item{target}{Type: list. A list to perform \code{cbindlist} on, with vectors only.}
}
\value{
A data.table based on \code{target}.
}
\description{
This function attempts to replicate \code{rbindlist} for columns.
}
\examples{
library(data.table)

# Create named junk list
mega_list <- list(a = rnorm(1000),
                  b = rnorm(1000),
                  c = rnorm(1000),
                  d = rnorm(1000),
                  e = rnorm(1000),
                  f = rnorm(1000),
                  g = rnorm(1000))

# Perform cbindlist, equivalent of rbindlist but for columns, accepting only vectors
my_dt <- cbindlist(mega_list)

# Check column name correctness/coverage
sum(colnames(my_dt) == c("a", "b", "c", "d", "e", "f", "g")) / 7 # 100\% coverage

}

