% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTree_Forest_pred_internals.R
\name{CRTree_Forest_pred_internals}
\alias{CRTree_Forest_pred_internals}
\title{Complete-Random Tree Forest Predictor (Deferred predictor) implementation in R}
\usage{
CRTree_Forest_pred_internals(model, data, folds = NULL, prediction = FALSE,
  multi_class = NULL, data_start = NULL, return_list = TRUE)
}
\arguments{
\item{model}{Type: list. A model trained by \code{CRTreeForest}.}

\item{data}{Type: data.table. A data to predict on. If passing training data, it will predict as if it was out of fold and you will overfit (so, use the list \code{train_preds} instead please).}

\item{folds}{Type: list. The folds as list for cross-validation if using the training data. Otherwise, leave \code{NULL}. Defaults to \code{NULL}.}

\item{prediction}{Type: logical. Whether the predictions of the forest ensemble are averaged. Set it to \code{FALSE} for debugging / feature engineering. Setting it to \code{TRUE} overrides \code{return_list}. Defaults to \code{FALSE}.}

\item{multi_class}{Type: numeric. How many classes you got. Set to 2 for binary classification, or regression cases. Set to \code{NULL} to let it try guessing by reading the \code{model}. Defaults to \code{NULL}.}

\item{data_start}{Type: vector of numeric. The initial prediction labels. Set to \code{NULL} if you do not know what you are doing. Defaults to \code{NULL}.}

\item{return_list}{Type: logical. Whether lists should be returned instead of concatenated frames for predictions. Defaults to \code{TRUE}.}
}
\value{
A data.table or a list based on \code{data} predicted using \code{model}.
}
\description{
This function attempts to predict from Complete-Random Tree Forests using xgboost. Requesting predictions form \code{CRTreeForest} should be done using \code{CRTreeForest_pred}.
}
\details{
For implementation details of Cascade Forest / Complete-Random Tree Forest / Multi-Grained Scanning / Deep Forest, check this: \url{https://github.com/Microsoft/LightGBM/issues/331#issuecomment-283942390} by Laurae.
}

