% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_acc}
\alias{df_acc}
\title{Accuracy loss (Acc) (computation function, any size)}
\usage{
df_acc(preds, labels, threshold = 0.5)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}

\item{threshold}{The \code{threshold} for binary classification, if using binary classification. Defaults to \code{0.5}.}
}
\value{
The Accuracy.
}
\description{
This function computes the Accuracy loss (Acc) provided \code{preds} and \code{labels}, while handling multiclass problems.
}
\examples{
library(data.table)

# Binary classification problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- as.numeric(runif(n = 9*6*10) >= 0.5)
df_acc(my_preds, my_labels)

# Multiclass classification problem
my_preds <- data.table(matrix(dnorm(rnorm(n = 9*6*10)), nrow = 9*10))
my_labels <- rep(c(1, 2, 3, 4, 5, 1, 0, 2, 3), 10)
df_acc(my_preds, my_labels)

}

