% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_acc_bin}
\alias{df_acc_bin}
\title{Accuracy loss (Acc) (computation function, binary optimization)}
\usage{
df_acc_bin(preds, labels)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}
}
\value{
The optimized Accuracy.
}
\description{
This function computes the Accuracy loss (Acc) provided \code{preds} and \code{labels}, while optimizing for binary problems.
}
\examples{
library(data.table)

# Binary classification problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- as.numeric(runif(n = 9*6*10) >= 0.5)
df_acc_bin(my_preds, my_labels)

}

