% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_logloss}
\alias{df_logloss}
\title{Logarithmic Loss (Logloss) (computation function, any size)}
\usage{
df_logloss(preds, labels, eps = 1e-15)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}

\item{eps}{The threshold for predictions so you do not have infinite loss issued from an observation. Defaults to \code{1e-15}.}
}
\value{
The Logarithmic Loss.
}
\description{
This function computes the Logarithmic Loss (Logloss) provided \code{preds} and \code{labels}, while handling multiclass problems.
}
\examples{
library(data.table)

# Binary classification problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- as.numeric(runif(n = 9*6*10) >= 0.5)
df_logloss(my_preds, my_labels)

# Multiclass classification problem
my_preds <- data.table(matrix(dnorm(rnorm(n = 9*6*10)), nrow = 9*10))
my_labels <- rep(c(1, 2, 3, 4, 5, 1, 0, 2, 3), 10)
df_logloss(my_preds, my_labels)

}

