% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_mape}
\alias{df_mape}
\title{Mean Absolute Percentage Error (MAPE) (computation function, single vector)}
\usage{
df_mape(preds, labels)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}
}
\value{
The Mean Absolute Percentage Error.
}
\description{
This function computes the Mean Absolute Percentage Error loss (MAPE) provided \code{preds} and \code{labels}, but cannot handle multiclass problems.
}
\examples{
library(data.table)

# Regression problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- runif(n = 9*6*10)
df_mape(my_preds, my_labels)

}

