% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_mce}
\alias{df_mce}
\title{Mean Cubic Error (MCE) (computation function, any size)}
\usage{
df_mce(preds, labels)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}
}
\value{
The Mean Cubic Error.
}
\description{
This function computes the Mean Cubic Error loss (MCE) provided \code{preds} and \code{labels}, while handling multiclass problems.
}
\examples{
library(data.table)

# Binary classification problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- as.numeric(runif(n = 9*6*10) >= 0.5)
df_mce(my_preds, my_labels)

# Multiclass classification problem
my_preds <- data.table(matrix(dnorm(rnorm(n = 9*6*10)), nrow = 9*10))
my_labels <- rep(c(1, 2, 3, 4, 5, 1, 0, 2, 3), 10)
df_mce(my_preds, my_labels)

}

