% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_medae}
\alias{df_medae}
\title{Median Absolute Error (MedAE) (computation function, single vector)}
\usage{
df_medae(preds, labels)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}
}
\value{
The Median Absolute Error.
}
\description{
This function computes the Median Absolute Error loss (MedAE) provided \code{preds} and \code{labels}, but cannot handle multiclass problems.
}
\examples{
library(data.table)

# Regression problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- runif(n = 9*6*10)
df_medae(my_preds, my_labels)

}

