% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_loss.R
\name{df_r}
\alias{df_r}
\title{Pearson Coefficient of Correlation (R) (computation function, any size)}
\usage{
df_r(preds, labels)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{labels}{The \code{labels}.}
}
\value{
The Pearson Coefficient of Correlation.
}
\description{
This function computes the Pearson Coefficient of Correlation loss (R) provided \code{preds} and \code{labels}, while handling multiclass problems (mean).
}
\examples{
library(data.table)

# Regression problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- runif(n = 9*6*10)
df_r(my_preds, my_labels)

# Binary classification problem
my_preds <- dnorm(rnorm(n = 9*6*10))
my_labels <- as.numeric(runif(n = 9*6*10) >= 0.5)
df_r(my_preds, my_labels)

# Multiclass classification problem
my_preds <- data.table(matrix(dnorm(rnorm(n = 9*6*10)), nrow = 9*10))
my_labels <- rep(c(1, 2, 3, 4, 5, 1, 0, 2, 3), 10)
df_r(my_preds, my_labels)

}

