% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT2mat.R
\name{DT2mat}
\alias{DT2mat}
\title{data.table to matrix}
\usage{
DT2mat(DT, low_mem = FALSE, collect = 0, silent = TRUE)
}
\arguments{
\item{DT}{Type: data.table (or a data.frame, partially supported). The data.table to coerce to matrix.}

\item{low_mem}{Type: boolean. Unallows \code{DT} twice in memory by modifying \code{DT} in place. (WARNING: empties your \code{DT}) to save memory when set to \code{TRUE}. Setting it to \code{FALSE} allow \code{DT} to reside twice in memory, therefore memory usage increases. Defaults to \code{FALSE}.}

\item{collect}{Type: integer. Forces a garbage collect every \code{collect} iterations to clear up memory. Setting this to \code{1} along with \code{low_mem} = \code{TRUE} leads to the lowest possible memory usage one can ever get to merge two data.tables. It also prints verbose information about the process everytime it garbage collects. Setting this to \code{0} leads to no garbage collect. Lower values increases the time required to bind the data.tables. Defauls to \code{0}.}

\item{silent}{Type: boolean. Force silence during garbage collection iterations at no speed cost. Defaults to \code{TRUE}.}
}
\value{
A matrix.
}
\description{
This function coerces a data.table to a matrix using the least possible memory.
}
\details{
Warning: \code{DT} is a pointer only and is directly modified.
}
\examples{
library(data.table)
df <- data.table(matrix(1:50000000, nrow = 50000))
mat <- DT2mat(df, low_mem = TRUE, collect = 50, silent = FALSE)

}

