% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastROC.R
\name{FastROC}
\alias{FastROC}
\title{Fast AUROC (AUC, ROC) computation}
\usage{
FastROC(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
The AUC.
}
\description{
This function computes the AUROC fast.
}

