% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_acc.R
\name{get.max_acc}
\alias{get.max_acc}
\title{Maximum binary accuracy}
\usage{
get.max_acc(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the maximum accuracy for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to maximize the binary accuracy. A pair of values is returned.
}

