% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_f1.R
\name{get.max_f1}
\alias{get.max_f1}
\title{Maximum F1 Score (Precision with Sensitivity harmonic mean)}
\usage{
get.max_f1(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the maximum Maximum F1 Score (Precision with Sensitivity harmonic mean) for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to maximize the Maximum F1 Score (Precision with Sensitivity harmonic mean). A pair of values is returned.
}

