% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_fallout.R
\name{get.max_fallout}
\alias{get.max_fallout}
\title{Minimum Fall-Out (False Positive Rate)}
\usage{
get.max_fallout(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the minimum Fall-Out (False Positive Rate) for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to minimize the Fall-Out (False Positive Rate). A pair of values is returned.
}

