% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_kappa.R
\name{get.max_kappa}
\alias{get.max_kappa}
\title{Maximum Kappa statistic}
\usage{
get.max_kappa(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the maximum Kappa statistic for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to maximize the Kappa statistic. A pair of values is returned.
}

