% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_mcc.R
\name{get.max_mcc}
\alias{get.max_mcc}
\title{Maximum Matthews Correlation Coefficient}
\usage{
get.max_mcc(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the maximum Matthews Correlation Coefficient for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to maximize the Matthews Correlation Coefficient. A pair of values is returned.
}

