% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_sensitivity.R
\name{get.max_sensitivity}
\alias{get.max_sensitivity}
\title{Maximum Sensitivity (True Positive Rate)}
\usage{
get.max_sensitivity(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the maximum Sensitivity (True Positive Rate) for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to maximize the Sensitivity (True Positive Rate). A pair of values is returned.
}

