% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max_specificity.R
\name{get.max_specificity}
\alias{get.max_specificity}
\title{Maximum Specificity (True Negative Rate)}
\usage{
get.max_specificity(preds, labels)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}
}
\value{
A two element vector containing the maximum Specificity (True Negative Rate) for binary data, and the threshold used.
}
\description{
This function allows to use a custom thresholding method to maximize the Specificity (True Negative Rate). A pair of values is returned.
}

