% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{ggplot multiple plot per page}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{Type: ggplot. Input any ggplot you want.}

\item{ncol}{Type: integer. The number of columns in your final page. Defaults to \code{"length(list(...))"}.}

\item{nrow}{Type: itneger The number of rows in your final page. Defaults to \code{1}.}

\item{position}{Type: character. Where to put the legend. Defaults to \code{c("bottom", "right")}}
}
\value{
The requested ggplots in the same page.
}
\description{
This function allows to plot multiple ggplot plots in the same page while keepin the same legend.
}
\examples{
\dontrun{
dataset(mtcars)
p1 <- ggplot(mtcars, aes(x = mpg, y = as.factor(cyl), fill = vs)) + geom_point()
p2 <- ggplot(mtcars, aes(x = mpg, y = as.factor(cyl), fill = vs)) + geom_boxplot()
grid_arrange_shared_legend(p1, p2)
}

}

