% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.eda_d3js.R
\name{interactive.eda_d3js}
\alias{interactive.eda_d3js}
\title{Interactive Dashboard for Exploratory Data Analysis (ggplot)}
\usage{
interactive.eda_d3js(data, type = "scatterChart",
  dep_var = colnames(data)[1], indep_var = "None", group_var = "None",
  filter_var = "None", plot_width = 1500, plot_height = 820,
  f_back = "red", side_width = 300)
}
\arguments{
\item{data}{Type: name reference to a data.table (preferred) or data.frame. The data you want to explore. Using a data.table increases the processing speed.}

\item{type}{Type: character. The type of plot requested to d3js nvds. You may choose between \code{"lineChart"}, \code{"scatterChart"}, \code{"stackedAreaChart"}, \code{"discreteBarChart"}, \code{"multiBarChart"}, \code{"multiBarHorizontalChart"}, \code{"linePlusBarChart"}, \code{"cumulativeLineChart"}, \code{"lineWithFocusChart"}, \code{"pieChart"}, \code{"bulletChart"}. Defaults to \code{"scatterChart"}.}

\item{dep_var}{Type: character (currently overriden). The dependent variable to plot. Defaults to \code{colnames(data)[1]}, a reference to the first column of \code{data}.}

\item{indep_var}{Type: character (currently overriden). The independent variable to plot. Defaults to \code{"None"}.}

\item{group_var}{Type: character (currently overriden). The grouping variable to plot. Defaults to \code{"None"}.}

\item{filter_var}{Type: character (feature currently broken). The filtering variable to plot. Defaults to \code{"None"}.}

\item{plot_width}{Type: numeric. The width for the plot. Defaults to \code{1500}, which fits nicely Full HD screens (1920 vertical pixels).}

\item{plot_height}{Type: numeric. The height for the plot. Defaults to \code{820}, which fits nicely Full HD screens (1080 vertical pixels).}

\item{f_back}{Type: character. A background color character for the header. Defaults to \code{"red"}.}

\item{side_width}{Type: numeric. The width of the sidebar containing variable names. Defaults to \code{300}.}
}
\value{
Nothing
}
\description{
This function runs an interactive dashboard which allows to explore data via 3djs and NVD3. Unlike its counterpart \code{interactive.eda_ggplot}, using 3djs is not efficient for large data (10K+ observations) and automated smart plotting is not available. In addition, the filtering method is currently broken.
}
\details{
Plotting is done using \code{rCharts}, which must be loaded before running this function. There are issues if you open this in Internet Explorer 9 or under.

The colors (header: \code{f_back}) allowed are the following:

\describe{
  \item{blue}{blue color}
  \item{black}{black color}
  \item{purple}{purple color}
  \item{green}{green color}
  \item{red}{red color}
  \item{yellow}{yellow color}
}
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(rCharts)
library(datasets)
hair_eye <- as.data.frame(HairEyeColor)
interactive.eda_d3js(data = "hair_eye",
                     type = "multiBarChart",
                     dep_var = "Hair", #ignored
                     indep_var = "Freq", #ignored
                     group_var = "Sex", #ignored
                     plot_width = 600,
                     plot_height = 420,
                     f_back = "red",
                     side_width = 300)
}

}

