% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.eda_ggplot.R
\name{interactive.eda_ggplot}
\alias{interactive.eda_ggplot}
\title{Interactive Dashboard for Exploratory Data Analysis (ggplot)}
\usage{
interactive.eda_ggplot(data, p_back = "red", f_back = "red",
  side_width = 300, max_height = 820)
}
\arguments{
\item{data}{Type: data.table (preferred) or data.frame. The data you want to explore. Using a data.table allows to avoid copying in-memory when switching datasets.}

\item{p_back}{Type: character. A background color character for the plot frame. Defaults to \code{"red"}.}

\item{f_back}{Type: character. A background color character for the header. Defaults to \code{"red"}.}

\item{side_width}{Type: numeric. The width of the sidebar containing variable names. Defaults to \code{300}.}

\item{max_height}{Type: numeric. The maximum height for the plots. Defaults to \code{820}, which fits nicely Full HD screens (820 vertical pixels).}
}
\value{
Nothing
}
\description{
This function runs an interactive dashboard which allows to explore data in a WYSIWYG (What You See Is What You Get). It uses automated variable detection along with smart automated plotting, which eases the task for data exploration.
}
\details{
Plotting is done using \code{plotluck}, which must be loaded along \code{ggplot2}.

The colors (\code{p_back}) allowed are the following:

\describe{
  \item{red}{red color}
  \item{yellow}{yellow color}
  \item{aqua}{aqua color}
  \item{blue}{blue color}
  \item{light-blue}{light-blue color}
  \item{green}{green color}
  \item{navy}{navy color}
  \item{teal}{teal color}
  \item{olive}{olive color}
  \item{lime}{lime color}
  \item{orange}{orange color}
  \item{fuchsia}{fuchsia color}
  \item{purple}{purple color}
  \item{maroon}{maroon color}
  \item{black}{black color}
}

The colors (header: \code{f_back}) allowed are the following:

\describe{
  \item{blue}{blue color}
  \item{black}{black color}
  \item{purple}{purple color}
  \item{green}{green color}
  \item{red}{red color}
  \item{yellow}{yellow color}
}
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(ggplot2)
library(plotluck)
library(datasets)
data(faithful)
interactive.eda_ggplot(data = faithful,
                       p_back = "red",
                       f_back = "red",
                       side_width = 300,
                       max_height = 820)
}

}

