% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.eda_plotly.R
\name{interactive.eda_plotly}
\alias{interactive.eda_plotly}
\title{Interactive Dashboard for Exploratory Data Analysis (Plotly)}
\usage{
interactive.eda_plotly(data, type = "scatter", plot_width = 1500,
  plot_height = 820, f_back = "red", side_width = 300)
}
\arguments{
\item{data}{Type: name reference to a data.table (preferred) or data.frame. The data you want to explore. Using a data.table increases the processing speed.}

\item{type}{Type: character. The type of plot to preselect (but it does nothing itself). You may choose between \code{"scatter"}, \code{"bar"}, \code{"pie"}, \code{"histogram"}, \code{"histogram2d"}, \code{"box"}, \code{"contour"}, \code{"heatmap"}, \code{"polar"}, \code{"scatter3d"}, \code{"surface"}, \code{"geom_density"}, \code{"geom_density_2d"}, \code{"stat_density_2d"}, \code{"geom_bar"}, \code{"geom_point"}, and \code{"geom_boxplot"}. Defaults to \code{"scatter"}.}

\item{plot_width}{Type: numeric. The width for the plot. Defaults to \code{1500}, which fits nicely Full HD screens (1920 vertical pixels).}

\item{plot_height}{Type: numeric. The height for the plot. Defaults to \code{820}, which fits nicely Full HD screens (1080 vertical pixels).}

\item{f_back}{Type: character. A background color character for the header. Defaults to \code{"red"}.}

\item{side_width}{Type: numeric. The width of the sidebar containing variable names. Defaults to \code{300}.}
}
\value{
Nothing
}
\description{
This function runs an interactive dashboard which allows to explore data via 3djs and Plotly (using Plotly or Plotly+ggplot2). Unlike its counterpart \code{interactive.eda_ggplot}, using 3djs is not efficient for large data (10K+ observations) and automated smart plotting is not available. You should use this function over \code{interactive.eda_3djs}, which performs poorly compared to its plotly counterpart. When specific variables are mandatory, the interactive dashboard will warn you with a star preceding the element which is mandatory to be used. In addition, make sure your variables are in the right formats, as columns may be turned into factors (this is true for facetting, coloring...).
}
\details{
Plotting is done using \code{plotly}, which must be loaded before running this function. You must import data as \code{data.table} if you want maximum performance, although you will probably not notice any difference unless you are playing with millions of rows. For kernel estimation, you need the \code{MASS} package. There are issues if you open this in Internet Explorer 9 or under.

The colors (header: \code{f_back}) allowed are the following:

\describe{
  \item{blue}{blue color}
  \item{black}{black color}
  \item{purple}{purple color}
  \item{green}{green color}
  \item{red}{red color}
  \item{yellow}{yellow color}
}
}
\examples{
\dontrun{
library(shiny)
library(shinydashboard)
library(MASS)
library(plotly)
library(datasets)
hair_eye <- as.data.frame(HairEyeColor)
interactive.eda_plotly(data = hair_eye,
                       type = "scatter",
                       plot_width = 600,
                       plot_height = 420,
                       f_back = "red",
                       side_width = 300)
}

}

