% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.eda_RColorBrewer.R
\name{interactive.eda_RColorBrewer}
\alias{interactive.eda_RColorBrewer}
\title{Interactive Dashboard for Finding the Perfect Color Brewer Palette}
\usage{
interactive.eda_RColorBrewer(f_back = "red", side_width = 300,
  plot_height = 360)
}
\arguments{
\item{f_back}{Type: character. A background color character for the header. Defaults to \code{"red"}.}

\item{side_width}{Type: numeric. The width of the sidebar containing variable names. Defaults to \code{300}.}

\item{plot_height}{Type: numeric. The height for the plot. Defaults to \code{360}, which fits nicely Full HD screens (1080 vertical pixels).}
}
\value{
A bandwidth on a scale suitable for the \code{width} argument of \code{density}.
}
\description{
This interactive dashboard allows the user to find the perfect palette to use from Color Brewer palettes. In addition, an unlimited amount of values is provided, instead of the practical limitations in the original Color Brewer palette. You can also request RGB values, and copy and paste immediately results from text (a vector ready to be pasted in R is provided back as text, along with the evolution of the colors over the n-th color interpolation).
}
\details{
The colors (header: \code{f_back}) allowed are the following:

\describe{
  \item{blue}{blue color}
  \item{black}{black color}
  \item{purple}{purple color}
  \item{green}{green color}
  \item{red}{red color}
  \item{yellow}{yellow color}
}
}
\examples{
\dontrun{
interactive.eda_RColorBrewer()
}

}

