% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.SymbolicLoss.R
\name{interactive.SymbolicLoss}
\alias{interactive.SymbolicLoss}
\title{Interactive Dashboard for Symbolic Gradient/Hessian Loss Behavior Exploration}
\usage{
interactive.SymbolicLoss(f1 = "(x, y) {(x - y) ^ 2}",
  f2 = "(x, y) {(y * log(x) + (1 - y) * log(1 - x))}",
  f3 = "(x, y) {(x - y * log(x)) + (y * log(y) - y)}",
  f4 = "(x, y) {(y - x) * log(y / x)}", f1_init = c(-20, 20, 50, 0),
  f2_init = c(0.01, 0.99, 50, 1), f3_init = c(0, 100, 100, 20),
  f4_init = c(0, 100, 100, 20), f1_back = "yellow", f2_back = "aqua",
  f3_back = "olive", f4_back = "purple", f_back = "red", type = "o",
  max_height = 580)
}
\arguments{
\item{f1}{Type: character. A string describing the first loss function, with potentially multiple arguments. Requires at least \code{(x, y)} arguments. Defaults to \code{"(x, y) {(x - y) ^ 2}"}.}

\item{f2}{Type: character. A string describing the second loss function, with potentially multiple arguments. Requires at least \code{(x, y)} arguments. Defaults to \code{"(x, y) {(y * log(x) + (1 - y) * log(1 - x))}"}.}

\item{f3}{Type: character. A string describing the third loss function, with potentially multiple arguments. Requires at least \code{(x, y)} arguments. Defaults to \code{"(x, y) {(x - y * log(x)) + (y * log(y) - y)}"}.}

\item{f4}{Type: character. A string describing the fourth loss function, with potentially multiple arguments. Requires at least \code{(x, y)} arguments. Defaults to \code{"(x, y) {(y - x) * log(y / x)}"}.}

\item{f1_init}{Type: numeric vector of length 4. A vector containing sequentially the minimum, the maximum, the number of points, and the y value for the plots of the first loss function. Defaults to \code{c(-20, 20, 50, 0),}.}

\item{f2_init}{Type: numeric vector of length 4. A vector containing sequentially the minimum, the maximum, the number of points, and the y value for the plots of the second loss function. Defaults to \code{c(0.01, 0.99, 50, 1)}.}

\item{f3_init}{Type: numeric vector of length 4. A vector containing sequentially the minimum, the maximum, the number of points, and the y value for the plots of the third loss function. Defaults to \code{c(0, 100, 100, 20)}.}

\item{f4_init}{Type: numeric vector of length 4. A vector containing sequentially the minimum, the maximum, the number of points, and the y value for the plots of the fourth loss function. Defaults to \code{c(0, 100, 100, 20)}.}

\item{f1_back}{Type: character. A background color character for the first function. Defaults to \code{"yellow"}.}

\item{f2_back}{Type: character. A background color character for the second function. Defaults to \code{"aqua"}.}

\item{f3_back}{Type: character. A background color character for the third function. Defaults to \code{"olive"}.}

\item{f4_back}{Type: character. A background color character for the fourth function. Defaults to \code{"purple"}.}

\item{f_back}{Type: character. A background color character for the header. Defaults to \code{"red"}.}

\item{type}{Type: character. The type of plot to use for plots. \code{"p"} for points, \code{"l"} for lines, \code{"b"} for points+line, \code{"c"} for line without points, \code{"o"} for overplotted (points+line overlapping), \code{"h"} for high-density vertical lines (histogram-like), \code{"s"} for optimistic stair steps, \code{"S"} for pessimistic stair steps, \code{"n"} to plot nothing. Defaults to \code{"o"} for overplotted.}

\item{max_height}{Type: numeric. The maximum height for the plots. Defaults to \code{580}, which fits nicely Full HD screens (1080 vertical pixels).}
}
\value{
Nothing
}
\description{
This function runs an interactive dashboard which computes the 1st and 2nd symbolic derivatives of the loss function (gradient/hessian) provided, for up to 4 functions simultaneously without interrupting if you input a bad function.
}
\details{
This function cannot handle any type of input. It cannot handle sums or loops in the function code. It handles the following, in the alphabetic order:

\describe{
  \item{\*}{Multiplication}
  \item{/}{Division}
  \item{^}{Power}
  \item{abs}{Absolute value function}
  \item{acos}{Arcosine function}
  \item{acosh}{Hyperbolic Arcosine function}
  \item{asin}{Arsine function}
  \item{asinh}{Hyperbolic Arcsine function}
  \item{atan}{Arctangent function}
  \item{atan2}{Arctangent angle function between the x-axis and the vector from the origin (x,y), atan=y/x if x>0 and y>0}
  \item{atanh}{Hyperbolic Arctangent function}
  \item{besselI}{Modified Bessel function of the first kind}
  \item{besselJ}{Bessel function of the first kind}
  \item{besselK}{Modified Bessel function of the second kind}
  \item{besselY}{Sphereical Bessel function}
  \item{beta}{Beta function (Eulerian integral of the first kind)}
  \item{cos}{Cosine function}
  \item{cosh}{Hyperbolic cosine function}
  \item{cospi}{Cosine function with argument multiplicand pi}
  \item{dbinom}{Density binomial function}
  \item{digamma}{First derivative of the logarithm of the gamma function}
  \item{dnorm}{Density normal function}
  \item{exp}{Exponential function}
  \item{expm1}{Exponential function minus 1}
  \item{gamma}{Gamma function (Mellin transform of the negative exponential function)}
  \item{lbeta}{Natural logarithm of the beta function}
  \item{lgamma}{Natural logarithm of the absolute value of the gamma function}
  \item{log}{Natural (e) logarithm function}
  \item{log10}{Common (10) logarithm function}
  \item{log1p}{Natural (e) logarithm function with 1 added to the argument}
  \item{log2}{Binary (2) logarithm function}
  \item{logb}{Logarithm function of base b (base)}
  \item{pnorm}{Normal distribution function}
  \item{psigamma}{Polygamma function (degree specified by deriv)}
  \item{rep.int}{Replicate "times" elements of vectors and lists}
  \item{rep_len}{Replicate "length.out" elements of vectors and lists}
  \item{sign}{Sign function}
  \item{sin}{Sine function}
  \item{sinh}{Hyperbolic sine function}
  \item{sinpi}{Sine function with argument multiplicand pi}
  \item{sqrt}{Square root function}
  \item{tan}{Tangent function}
  \item{tanh}{Hyperbolic tangent function}
  \item{tanpi}{Tangent function with argument multiplicand pi}
  \item{trigamma}{Second derivative of the logarithm of the gamma function}
}

The colors (non header: \code{f1_back}, \code{f2_back}, \code{f3_back}, \code{f4_back}) allowed are the following:

\describe{
  \item{red}{red color}
  \item{yellow}{yellow color}
  \item{aqua}{aqua color}
  \item{blue}{blue color}
  \item{light-blue}{light-blue color}
  \item{green}{green color}
  \item{navy}{navy color}
  \item{teal}{teal color}
  \item{olive}{olive color}
  \item{lime}{lime color}
  \item{orange}{orange color}
  \item{fuchsia}{fuchsia color}
  \item{purple}{purple color}
  \item{maroon}{maroon color}
  \item{black}{black color}
}

The colors (header: \code{f_back}) allowed are the following:

\describe{
  \item{blue}{blue color}
  \item{black}{black color}
  \item{purple}{purple color}
  \item{green}{green color}
  \item{red}{red color}
  \item{yellow}{yellow color}
}
}
\examples{
\dontrun{
interactive.SymbolicLoss(f1 = "(x, y) {(x - y) ^ 2}",
                         f2 = "(x, y) {(y * log(x) + (1 - y) * log(1 - x))}",
                         f3 = "(x, y) {(x - y * log(x)) + (y * log(y) - y)}",
                         f4 = "(x, y) {(y - x) * log(y / x)}",
                         f1_init = c(-20, 20, 50, 0),
                         f2_init = c(0.01, 0.99, 50, 1),
                         f3_init = c(0, 100, 100, 20),
                         f4_init = c(0, 100, 100, 20),
                         f1_back = "yellow",
                         f2_back = "aqua",
                         f3_back = "olive",
                         f4_back = "purple",
                         f_back = "red",
                         type = "o",
                         max_height = 580)
}

}

