% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel2d_est.R
\name{kernel2d_est}
\alias{kernel2d_est}
\title{MASS' Two-Dimensional Kernel Density Estimation}
\usage{
kernel2d_est(x, y, h, n = 25, lims = c(range(x), range(y)))
}
\arguments{
\item{x}{Type: numeric vector. x coordinate of data.}

\item{y}{Type: numeric vector. y coordinate of data.}

\item{h}{Type: numeric vector. Vector of bandwidths for x and y directions. Defaults to normal reference bandwidth. A scalar value will be taken to apply to both directions.}

\item{n}{Type: numeric vector. Number of grid points in each direction. Can be scalar or a length-2 integer vector.}

\item{lims}{Type: numeric vector. The limits of the rectangle covered by the grid as \code{c(xl, xu, yl, yu)}.}
}
\value{
A list of three elements.
}
\description{
Two-dimensional kernel density estimation with an axis-aligned bivariate normal kernel, evaluated on a square grid. Originally from \code{MASS} package, this function was copied to avoid CRAN notes.
}
\details{
Return a list of three components:

\describe{
  \item{x}{The x coordinates of the grid points, vector of length n}
  \item{y}{The y coordinates of the grid points, vector of length n}
  \item{z}{An n[1] by n[2] matrix of the estimated density: rows correspond to the value of x, columns to the value of y.}
}
}
\examples{
\dontrun{
#?MASS::kde2d
}

}

