% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laurae.R
\docType{package}
\name{Laurae-package}
\alias{Laurae}
\alias{Laurae-package}
\title{Laurae's package for (very) advanced Data Science for R}
\description{
This pacakge is meant to be used for doing advanced statistics, visualization, and modeling.
}

