% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LauraeML_utils.R
\name{LauraeML_utils.badlog}
\alias{LauraeML_utils.badlog}
\title{Laurae's Machine Learning Utility: bad input logger}
\usage{
LauraeML_utils.badlog(mobile, logging, x, y, score = NA)
}
\arguments{
\item{mobile}{Type: environment. The environment passed from \code{LauraeML}.}

\item{logging}{Type: character. The \code{logging} parameter passed from \code{LauraeML} (where to store log file).}

\item{x}{Type: vector (numeric). The hyperparameters to use passed from the trainer.}

\item{y}{Type: vector (numeric). The features to use, as binary format (0 for not using, 1 for using) passed from the trainer.}

\item{score}{Type: numeric. The score to optimize.}
}
\description{
This function is a helper function to do the logging when a bad input is existing for \code{LauraeML} during a training iteration of the optimizer, specifically used when you have no features during training.
}
\examples{
\dontrun{
# What if we have no feature selected?
if (sum(y) == 0) {

  # Logging specific
  LauraeML_utils.badlog(logging, x, y, score = NA)
  
  # Last, we return an absurd score which is so high
  # you would rather have a random model than this 0-feature model
  # 
  # This iteration will be ignored by the optimizer if
  # it does not belong to the elite proportion of the optimization iteration
  # which should be obviously true
  return(LauraeML_utils.badinput(maximize,
                                 score = 9999999))
}
}

}

