% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LauraeML_utils.R
\name{LauraeML_utils.badscore}
\alias{LauraeML_utils.badscore}
\title{Laurae's Machine Learning Utility: bad input score}
\usage{
LauraeML_utils.badscore(maximize, score = 9999999)
}
\arguments{
\item{maximize}{Type: boolean. The \code{maximize} parameter passed from \code{LauraeML} (whether to maximize or not the metric).}

\item{score}{Type: numeric. The score to optimize.}
}
\value{
A so bad input score you better avoid such thing.
}
\description{
This function is a helper function to return a bad input score for \code{LauraeML} during a training iteration of the optimizer, specifically used when you have no features during training.
}
\examples{
\dontrun{
# What if we have no feature selected?
if (sum(y) == 0) {

  # Logging specific
  LauraeML_utils.badlog(logging, x, y, score = NA)
  
  # Last, we return an absurd score which is so high
  # you would rather have a random model than this 0-feature model
  # 
  # This iteration will be ignored by the optimizer if
  # it does not belong to the elite proportion of the optimization iteration
  # which should be obviously true
  return(LauraeML_utils.badinput(maximize,
                                 score = 9999999))
}
}

}

