% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LauraeML_utils.R
\name{LauraeML_utils.feat_sel}
\alias{LauraeML_utils.feat_sel}
\title{Laurae's Machine Learning Utility: subset features to select during training}
\usage{
LauraeML_utils.feat_sel(data, y)
}
\arguments{
\item{data}{Type: data.table. The data features. Comes from \code{LauraeML}.}

\item{y}{Type: vector (numeric). The features to use, as binary format (0 for not using, 1 for using) passed from the trainer.}
}
\description{
This function is a helper function to do the feature sampling for \code{LauraeML} during a training iteration of the optimizer.
}
\examples{
\dontrun{
mini_data <- LauraeML_utils.feat_sel(data, y)
}

}

