% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LauraeML_utils.R
\name{LauraeML_utils.xgb_data}
\alias{LauraeML_utils.xgb_data}
\title{Laurae's Machine Learning Utility: create xgboost dataset}
\usage{
LauraeML_utils.xgb_data(data, fold, label, is_train)
}
\arguments{
\item{data}{Type: data.table. The data features. Comes from \code{LauraeML}.}

\item{fold}{Type: vector (numeric). The observation rows to keep/remove. Comes from \code{LauraeML}.}

\item{label}{Type: vector (numeric). The labels. Comes from \code{LauraeML}.}

\item{is_train}{Type: logical. Whether the fold refers to testing data (\code{FALSE}) or training data (\code{TRUE}).}
}
\description{
This function is a helper function to do create a xgboost dataset for \code{LauraeML} during a training iteration of the optimizer.
}
\examples{
\dontrun{
for (i in 1:5) {
  # First we create the training data
  temp_train <- LauraeML_utils.xgb_data(data = mini_data,
                                        fold = folds[[i]],
                                        label = label,
                                        is_train = TRUE)

  # Second we create the testing data
  temp_test <- LauraeML_utils.xgb_data(data = mini_data,
                                       fold = folds[[i]],
                                       label = label,
                                       is_train = FALSE)
}
}

}

