% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbm.fi.R
\name{lgbm.fi}
\alias{lgbm.fi}
\title{LightGBM Feature Importance}
\usage{
lgbm.fi(workingdir, feature_names = NA, input_model = "LightGBM_model.txt",
  ntreelimit = 0)
}
\arguments{
\item{workingdir}{Type: character. The working directory of the model file.}

\item{feature_names}{Type: vector of characters. The names of the features, in the order they were fed to LightGBM. Returns column numbers if left as \code{NA}. Defaults to \code{NA}.}

\item{input_model}{Type: character. The name of the model file. Defaults to \code{"LightGBM_model.txt"}.}

\item{ntreelimit}{Type: integer. The number of trees to select, starting from the first tree. Defaults to \code{0}.}
}
\value{
Nothing (the data.table is transformed to a data.frame already)
}
\description{
This function allows to get the feature importance on a LightGBM model. The model file must be \code{"workingdir"}, where \code{"workingdir"} is the folder and \code{input_model} is the model file name.
}
\examples{
#None yet.

#@export
}

