% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbm.find.R
\name{lgbm.find}
\alias{lgbm.find}
\title{Find LightGBM Path}
\usage{
lgbm.find()
}
\value{
The path of the selected file which can be used for LightGBM (\code{lgbm_path} parameter).
}
\description{
This function opens a file browser which lets you select manually LightGBM.
}
\examples{
\dontrun{
lgbm.find()
}

}

