% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbm.metric.R
\name{lgbm.metric}
\alias{lgbm.metric}
\title{LightGBM Metric Output}
\usage{
lgbm.metric(model, metrics = TRUE, data.table = TRUE)
}
\arguments{
\item{model}{Type: list, vector of characters, character. If a list, it should come from a trained model. If a vector of characters, it must be the model text. If a character, it must be the exact file path to the log file. It cannot work without training with \code{verbose} set to \code{FALSE}.}

\item{metrics}{Type: boolean. Whether to return the metrics table (\code{TRUE}) or the best iteration (\code{FALSE}).}

\item{data.table}{Type: boolean. Whether to return a data.table (\code{TRUE}) or a data.frame (\code{FALSE}). Defaults to \code{TRUE}. Useless when \code{metrics} is set to \code{FALSE}.}
}
\value{
A data.table (or data.frame) with 3 to N columns, with at least: \code{c("Iteration", "Total_Time", "Round_Time")}.
}
\description{
This function allows to get the metric values from a LightGBM log.
}
\examples{
\dontrun{
# Get the metrics of the model trained (requires verbose = FALSE during training)
lgbm.metric(workingdir = trained[["Path"]],
log_name = trained[["Log"]],
metrics = TRUE,
data.table = TRUE)

# Get the best iteration of the model trained (requires verbose = FALSE during training)
lgbm.metric(workingdir = trained[["Path"]],
log_name = trained[["Log"]],
metrics = FALSE)
}

}

