% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbm.predict.R
\name{lgbm.predict}
\alias{lgbm.predict}
\title{LightGBM Prediction}
\usage{
lgbm.predict(model, x_pred = NA, y_pred = NA, data_has_label = FALSE,
  val_name = ifelse(is.list(model) & is.null(dim(x_pred)), model[["Test"]],
  "lgbm_test.csv"), input_model = ifelse(is.list(model), model[["Name"]],
  "lgbm_model.txt"), output_result = "lgbm_predict_result.txt",
  lgbm_path = ifelse(is.list(model), model[["lgbm"]], getwd()),
  workingdir = ifelse(is.list(model), model[["Path"]], getwd()),
  files_exist = TRUE, pred_conf = "lgbm_pred.conf",
  data.table = exists("data.table"), verbose = TRUE)
}
\arguments{
\item{model}{Type: list. The model file. If a character vector is provided, it is considered to be the model which is going to be saved as \code{input_model}. If a list is provided, it is used to setup to fetch the correct variables, which you can override by setting the arguments manually. If a single value is provided (like \code{NA}), then it is ignored and uses the other arguments to fetch the model locally.}

\item{x_pred}{Type: data.table (preferred), data.frame, or matrix. The validation features. Defaults to \code{NA}.}

\item{y_pred}{Type: vector. The validation labels. Defaults to \code{NA}.}

\item{data_has_label}{Type: boolean. Whether the data has labels or not. Do not modify this. Defaults to \code{FALSE}.}

\item{val_name}{Type: character. The file output name for the vaildation file. Defaults to \code{ifelse(is.list(model) & is.null(dim(x_pred)), model[["Test"]], 'lgbm_test.csv')}, which means "take the test file name if provided the model list and x_pred is left as is, else take "lgbm_test.csv".}

\item{input_model}{Type: character. The file name of the model. Defaults to \code{ifelse(is.list(model), model[["Name"]], 'lgbm_model.txt')}, which means "take the input model name if provided the model list, else take "lgbm_model.txt".}

\item{output_result}{Type: character. The output prediction file. Defaults to \code{'lgbm_predict_result.txt'}.}

\item{lgbm_path}{Type: character. Where is stored LightGBM? Include only the folder to it. Defaults to \code{ifelse(is.list(model), model[["File"]], getwd())}, which means "take the model LightGBM path if provided the model list, else take the default working directory".}

\item{workingdir}{Type: character. The working directory used for LightGBM. Defaults to \code{ifelse(is.list(model), model[["Path"]], getwd())}, which means "take the model working directory if provided the model list, else take the default working directory".}

\item{files_exist}{Type: boolean. Whether to NOT create CSV files for the prediction data, if already created. Defaults to \code{TRUE}.}

\item{pred_conf}{Type: character. The name of the pred_conf file for the model. Defaults to \code{'lgbm_pred.conf'}}

\item{data.table}{Type: boolean. Whether to use data.table to read data (returns a data.table). Defaults to \code{exists("data.table")}.}

\item{verbose}{Type: boolean. Whether to print to console verbose information. When FALSE, the printing is diverted to \code{"diverted_verbose.txt"}. Defaults to \code{TRUE}. Might not work when your lgbm_path has a space.}
}
\value{
The predictions.
}
\description{
This function allows to run predictions on provided data.
}
\examples{
#None yet.

}

