% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.predict.R
\name{lightgbm.predict}
\alias{lightgbm.predict}
\title{LightGBM Prediction}
\usage{
lightgbm.predict(model, x_val, y_val = NA, data_has_label = TRUE,
  val_name = "test.csv", input_model = "LightGBM_model.txt",
  output_result = "LightGBM_predict_result.txt",
  gbmpath = "/home/dba/KAGGLE/LightGBM", newx = TRUE,
  data.table = exists("data.table"))
}
\arguments{
\item{model}{Type: character.The working directory of the model.}

\item{x_val}{Type: data.table (preferred), data.frame, or matrix. The validation features.}

\item{y_val}{Type: vector. The validation labels. Defaults to \code{NULL}.}

\item{data_has_label}{Type: boolean. Whether the data has labels or not. Do not modify this. Defaults to \code{TRUE}.}

\item{val_name}{Type: character. The file output name for the vaildation file. Defaults to \code{"test.csv"}.}

\item{input_model}{Type: character. The file name of the model. Defaults to \code{'LightGBM_model.txt'}.}

\item{output_result}{Type: character. The output prediction file. Defaults to \code{'LightGBM_predict_result.txt'}.}

\item{gbmpath}{Type: character. Where is stored LightGBM? Include only the folder to it. Defaults to \code{'/home/dba/KAGGLE/LightGBM'}.}

\item{newx}{Type: boolean. Whether to NOT create CSV files for validation data, if already created. Defaults to \code{TRUE}.}

\item{data.table}{Type: boolean. Whether to use data.table to read data (returns a data.table). Defaults to \code{exists("data.table")}.}
}
\value{
The predictions.
}
\description{
This function allows to run predictions on provided data.
}
\examples{
None yet.

}

