% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLoss.R
\name{LogLoss}
\alias{LogLoss}
\title{Fast Logarithmic Loss (logloss) computation}
\usage{
LogLoss(preds, labels, eps = 1e-15)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{eps}{Type: numeric. The shrinkage between the 0 and 1's bounds.}
}
\value{
The Logarithmic Loss.
}
\description{
This function computes the Logarithmic Lost fast.
}

