% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_LL_grad}
\alias{loss_LL_grad}
\title{Loglikelihood Error (gradient function)}
\usage{
loss_LL_grad(y_pred, y_true)
}
\arguments{
\item{y_pred}{The \code{predictions}.}

\item{y_true}{The \code{labels}.}
}
\value{
The gradient of the Loglikelihood Error per value.
}
\description{
This function computes the Loglikelihood loss (logloss) gradient per value provided \code{preds} and \code{labels} probabilities.
}
\details{
Loss Formula : \eqn{- (y_true * log(y_pred) + (1 - y_true) * log(1 - y_pred))}

Gradient Formula : \eqn{(1 - y_true)/(1 - y_pred) - y_true/y_pred}

Hessian Formula : \eqn{(1 - y_true)/((1 - y_pred) * (1 - y_pred)) + y_true/(y_pred * y_pred)}
}

