% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_LL_math}
\alias{loss_LL_math}
\title{Loglikelihood Error (math function)}
\usage{
loss_LL_math(x, y)
}
\arguments{
\item{x}{The \code{predictions}.}

\item{y}{The \code{labels}.}
}
\value{
The Loglikelihood Error per value.
}
\description{
This function computes the Loglikelihood Error loss (logloss) per value provided \code{x, y} (preds, labels) probabilities.
}
\details{
Loss Formula : \eqn{- (y_true * log(y_pred) + (1 - y_true) * log(1 - y_pred))}

Gradient Formula : \eqn{(1 - y_true)/(1 - y_pred) - y_true/y_pred}

Hessian Formula : \eqn{(1 - y_true)/((1 - y_pred) * (1 - y_pred)) + y_true/(y_pred * y_pred)}
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_LL_math, xmin = 0, xmax = 1, y = rep(0, 21))
}

}

