% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_LL_xgb}
\alias{loss_LL_xgb}
\title{Loglikelihood Error (xgboost function)}
\usage{
loss_LL_xgb(preds, dtrain)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{dtrain}{The xgboost model.}
}
\value{
The gradient and the hessian of the Loglikelihood Error per value in a list.
}
\description{
This function computes for xgboost's \code{obj} function the Loglikelihood Error loss (logloss) gradient and hessian per value provided \code{preds} and \code{dtrain}.
}
\details{
Loss Formula : \eqn{- (y_true * log(y_pred) + (1 - y_true) * log(1 - y_pred))}

Gradient Formula : \eqn{(1 - y_true)/(1 - y_pred) - y_true/y_pred}

Hessian Formula : \eqn{(1 - y_true)/((1 - y_pred) * (1 - y_pred)) + y_true/(y_pred * y_pred)}
}

