% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_logloss}
\alias{loss_logloss}
\title{Loglikelihood Error (computation function)}
\usage{
loss_logloss(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The Loglikelihood Error per value.
}
\description{
This function computes the Loglikelihood Error loss (logloss) per value provided \code{preds} and \code{labels} probabilities.
}

