% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_logloss_grad}
\alias{loss_logloss_grad}
\title{Loglikelihood Error (gradient function)}
\usage{
loss_logloss_grad(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The gradient of the Loglikelihood Error per value.
}
\description{
This function computes the Loglikelihood loss (logloss) gradient per value provided \code{preds} and \code{labels} probabilities.
}

