% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_logloss_hess}
\alias{loss_logloss_hess}
\title{Loglikelihood Error (hessian function)}
\usage{
loss_logloss_hess(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The hessian of the Loglikelihood Error per value.
}
\description{
This function computes the MLoglikelihood Error loss (logloss) hessian per value provided \code{preds} and \code{labels} probabilities.
}

