% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_logloss_math}
\alias{loss_logloss_math}
\title{Loglikelihood Error (math function)}
\usage{
loss_logloss_math(x, y)
}
\arguments{
\item{x}{The \code{prediction}.}

\item{y}{The \code{label}.}
}
\value{
The Loglikelihood Error per value.
}
\description{
This function computes the Loglikelihood Error loss (logloss) per value provided \code{x, y} (preds, labels) probabilities.
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_logloss_math, fc_ref, xmin = 0, xmax = 1, y = rep(0, 21))
}

}

