% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_logloss_xgb}
\alias{loss_logloss_xgb}
\title{Loglikelihood Error (xgboost function)}
\usage{
loss_logloss_xgb(preds, dtrain)
}
\arguments{
\item{preds}{The predictions.}

\item{dtrain}{The labels.}
}
\value{
The gradient and the hessian of the Loglikelihood Error per value in a list.
}
\description{
This function computes for xgboost's \code{obj} function the Loglikelihood Error loss (logloss) gradient and hessian per value provided \code{preds} and \code{dtrain}.
}

