% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MAE}
\alias{loss_MAE}
\title{Mean Absolute Error (computation function)}
\usage{
loss_MAE(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The Squared Error per value.
}
\description{
This function computes the Mean Absolute Error loss (MAE) per value provided \code{preds} and \code{labels}.
}

