% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MAE_hess}
\alias{loss_MAE_hess}
\title{Mean Absolute Error (hessian function)}
\usage{
loss_MAE_hess(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The hessian of the Absolute Error per value.
}
\description{
This function computes the Mean Absolute Error loss (MAE) hessian per value provided \code{preds} and \code{labels}.
}

