% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MAE_math}
\alias{loss_MAE_math}
\title{Mean Absolute Error (symbolic function)}
\usage{
loss_MAE_math(x)
}
\arguments{
\item{x}{The difference between the \code{label} and the \code{prediction}.}
}
\value{
The Squared Error per value.
}
\description{
This function computes the Mean Absolute Error loss (MAE) per value provided \code{x} (preds - labels).
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_MAE_math, fc_ref)
}

}

