% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MCE}
\alias{loss_MCE}
\title{Mean Cubic Error (computation function)}
\usage{
loss_MCE(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The Cubic Error per value.
}
\description{
This function computes the Mean Cubic Error loss (MCE) per value provided \code{preds} and \code{labels}.
}

