% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MCE_grad}
\alias{loss_MCE_grad}
\title{Mean Cubic Error (gradient function)}
\usage{
loss_MCE_grad(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The gradient of the Cubic Error per value.
}
\description{
This function computes the Mean Cubic Error loss (MCE) gradient per value provided \code{preds} and \code{labels}.
}

