% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MSE}
\alias{loss_MSE}
\title{Mean Squared Error (computation function)}
\usage{
loss_MSE(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The Squared Error per value.
}
\description{
This function computes the Mean Squared Error loss (MSE) per value provided \code{preds} and \code{labels}.
}

