% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MSE_grad}
\alias{loss_MSE_grad}
\title{Mean Squared Error (gradient function)}
\usage{
loss_MSE_grad(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The gradient of the Squared Error per value.
}
\description{
This function computes the Mean Squared Error loss (MSE) gradient per value provided \code{preds} and \code{labels}.
}

