% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MSE_hess}
\alias{loss_MSE_hess}
\title{Mean Squared Error (hessian function)}
\usage{
loss_MSE_hess(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The hessian of the Squared Error per value.
}
\description{
This function computes the Mean Squared Error loss (MSE) hessian per value provided \code{preds} and \code{labels}.
}

