% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MSE_math}
\alias{loss_MSE_math}
\title{Mean Squared Error (symbolic function)}
\usage{
loss_MSE_math(x, y)
}
\arguments{
\item{x}{The \code{predictions}.}

\item{y}{The \code{labels}.}
}
\value{
The Squared Error per value.
}
\description{
This function computes the Mean Squared Error loss (MSE) per value provided \code{x} (preds - labels).
}
\details{
Supposing: \eqn{x = preds - labels}

Loss Formula : \eqn{x^2}

Gradient Formula : \eqn{2 * x}

Hessian Formula : \eqn{2}
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_MSE_math)
}

}

