% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_Poisson}
\alias{loss_Poisson}
\title{Laurae's Poisson Error (computation function)}
\usage{
loss_Poisson(y_pred, y_true)
}
\arguments{
\item{y_pred}{The \code{predictions}.}

\item{y_true}{The \code{labels}.}
}
\value{
The Laurae's Poisson Error per value.
}
\description{
This function computes the Laurae's Poisson Error loss per value provided \code{preds} and \code{labels} count.
}
\details{
This loss function is strictly positive, therefore defined in \code{\]0, +Inf\[}. It penalizes lower values more heavily, and as such is a good fit for typical problems requiring fine tuning when undercommitting on the predictions. The negative values are cancelled out to make the loss function positive, with \code{loss = 0} when \code{y_true = y_pred}. This loss function is experimental.

Loss Formula : \eqn{(y_pred - y_true * log(y_pred))}

Gradient Formula : \eqn{1 - y_true/y_pred}

Hessian Formula : \eqn{y_true/(y_pred * y_pred)}
}

