% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.R
\name{mean2}
\alias{mean2}
\title{Fast mean computation}
\usage{
mean2(x)
}
\arguments{
\item{x}{Type: unknown. The variable to perform \code{sum(x)/length(x)} on.}
}
\value{
The mean.
}
\description{
This function allows to perform \code{sum(x) / length(x)} in one shot, without performing any background check on the data. Beware of integer overflow!
}
\examples{
mean2(c(1:1000000))

}

