% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dep.feature.R
\name{partial_dep.feature}
\alias{partial_dep.feature}
\title{Partial Dependency, output analyzer}
\usage{
partial_dep.feature(grid_data, method = "emp", in_depth = FALSE)
}
\arguments{
\item{grid_data}{Type: data.table. A \code{partial_dep} \code{grid_exp} output.}

\item{method}{Type: character. The method to use to compute information theory -related values. Defaults to \code{"emp"}, see details for more information.}

\item{in_depth}{Type: logical. Whether to perform Partial Mann-Kendall test.}
}
\value{
Value statistics of the input for variability per column.
}
\description{
This function is a helper for analyzing partial dependency output. It uses information theory to assess the score of a predictor against the label.
}
\details{
There are multiple outputs, all information theory -related values (entropy, mutual information, synergy, total correlation) are provided with the empirical probability distribution by default. One can change them using the parmaeter \code{method}, but it is recommended to leave the default on unless you know what you are doing:

\describe{
  \item{"emp"}{(Default) Entropy of the empirical probability distribution.}
  \item{"mm"}{Miller-Madow asymptotic bias corrected empirical estimator.}
  \item{"shrink"}{Shrinkage estimate of the entropy of a Dirichlet probability distribution.}
  \item{"sg"}{Schurmann-Grassberger estimate of the entropy of a Dirichlet probability distribution.}
}

Use the function \code{infotheo::natstobits(value)} to convert from nats (\code{base e}) to bits (\code{base 2}).

Table \code{"Features"}:

\describe{
  \item{"Feature"}{Column name assessed.}
  \item{"Unique_Values"}{Count of unique values of the feature.}
  \item{"Entropy"}{Entropy of the feature.}
  \item{"Label_Mutual_Info"}{Mutual Information between Feature and Label.}
  \item{"Evolution_Mutual_Info"}{Mutual Information between Feature and Evolution.}
  \item{"p.Mann_Kendall"}{\code{in_depth = TRUE} p-value of the Partial Mann-Kendall (multivariate) test to detect non-parametric monotonic trends in potentially seasonal data. Low value means confidence in a trend.}
  \item{"p.Partial_Spearman"}{\code{in_depth = TRUE} p-value of the Partial Spearman Correlation trend test. Low value means confidence in a correlation greater than 0.}
}

Table \code{"Global"}:

\describe{
  \item{"Without_Feature"}{Column name NOT assessed.}
  \item{"Synergy"}{Synergy/Complementarity (inter information) provided by the data without the feature.}
  \item{"Total_Correlation"}{Total Correlation (multi information) provided by the data without the feature.}
}
}

