% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dep.plot.R
\name{partial_dep.plot}
\alias{partial_dep.plot}
\title{Partial Dependency, plotting function}
\usage{
partial_dep.plot(grid_data, backend = "tableplot", label_name = "Target",
  comparator_name = "Evolution", ...)
}
\arguments{
\item{grid_data}{Type: data.table. A \code{partial_dep} \code{grid_exp} output.}

\item{backend}{Type: logical. What type of backend for plotting to use. Check details for detailed description. Defaults to \code{tableplot}.}

\item{label_name}{Type: character. The label column name in \code{grid_data} when using \code{tableplot} or \code{lattice}.}

\item{comparator_name}{Type: character. The comparator column name in \code{grid_data} when using \code{tableplot} or \code{lattice}.}

\item{...}{other arguments to pass to to the plotting backend function.}
}
\value{
A plot with the requested backend.
}
\description{
This function is a helper to help plotting partial dependency plots using a provided grid via the specified backend.
}
\details{
For selecting the plotting backend, it depends on what you are trying to plot (from what you are working from). If you are using single observations (\code{partial_dep.obs}), you are provided the following backends:

\describe{
  \item{"tableplot"}{Tableplot is the best when it comes to plotting any type of large data. It is the default and the most appropriate for 99\% of cases, even if you have million of data points it will be blazing fast. Use this unless you have a rationale reason to use something else.}
  \item{"car"}{Car is the best when it comes to analyzing in depth the output in a scatter plot matrix, but it is extremely slow. Not recommended for more than 10k observations and 5 columns.}
  \item{"lattice"}{Lattice is used with parallel plots. Not recommended for more than 50k observations and 5 columns.}
  \item{"ggplot"}{ggplot is used for scatter plot matrix and correlation measurement. Not recommended for more than 20k observations and 5 columns.}
  \item{"plotly"}{Combines ggplot with Plotly for interactive graphics. Not recommended for more than 1k observations and 5 columns.}
  \item{"base"}{Base ships with R and thus is simple, but is slow. Not recommended for more than 50k observations and 5 columns.}
}

If you are using multiple observations (\code{partial_dep.obs_all}), you are provided the following backends:

\describe{
  \item{c("tableplot")}{Tableplot is used to output plot, but it is clearly not the recommended way of doing it. It is the default in case you have too many points.}
  \item{c("ggplot", "boxplot")}{ggplot is used to draw boxplots to check for distribution in boxplots, grouped by feature.}
  \item{c("ggplot", "point")}{ggplot is used to draw points and check for evolution, grouped both by feature and evolution.}
  \item{c("ggplot", "line")}{ggplot is used to draw points and lines and check for evolution, grouped both by feature and evolution.}
  \item{c("plotly", "boxplot")}{plotly + ggplot is used to draw boxplots to check for distribution in boxplots, grouped by feature, interactively.}
  \item{c("plotly", "point")}{plotly + ggplot is used to draw points and check for evolution, grouped both by feature and evolution, interactively.}
  \item{c("plotly", "line")}{plotly + ggplot is used to draw points and line and check for evolution, grouped both by feature and evolution, interactively.}
}
}
\examples{
\dontrun{
# Train you supervised machine learning model
# ...

# Prepare partial dependence content
my_grid <- partial_dep.obs(...)

# Plot partial dependence content
partial_dep.plot(my_grid$grid_exp, backend = "tableplot")
}


}

