% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_acc.R
\name{plotting.max_acc}
\alias{plotting.max_acc}
\title{Maximum binary accuracy plotting}
\usage{
plotting.max_acc(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their accuracy.
}
\description{
This function allows to use a custom thresholding method to maximize the binary accuracy. A data.table of values is returned.
}

