% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_f1.R
\name{plotting.max_f1}
\alias{plotting.max_f1}
\title{Maximum F1 Score (Precision with Sensitivity harmonic mean) plotting}
\usage{
plotting.max_f1(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their F1 Score (Precision with Sensitivity harmonic mean).
}
\description{
This function allows to use a custom thresholding method to maximize the F1 Score (Precision with Sensitivity harmonic mean). A data.table of values is returned.
}

