% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_fallout.R
\name{plotting.max_fallout}
\alias{plotting.max_fallout}
\title{Minimum Fall-Out (False Positive Rate) plotting}
\usage{
plotting.max_fallout(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their Fall-Out (False Positive Rate).
}
\description{
This function allows to use a custom thresholding method to minimize the Fall-Out (False Positive Rate). A data.table of values is returned.
}

