% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_mcc.R
\name{plotting.max_mcc}
\alias{plotting.max_mcc}
\title{Maximum Matthews Correlation Coefficient plotting}
\usage{
plotting.max_mcc(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their Matthews Correlation Coefficient.
}
\description{
This function allows to use a custom thresholding method to maximize the Matthews Correlation Coefficient. A data.table of values is returned.
}

