% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_missrate.R
\name{plotting.max_missrate}
\alias{plotting.max_missrate}
\title{Minimum Miss-Rate (False Negative Rate) plotting}
\usage{
plotting.max_missrate(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their Miss-Rate (False Negative Rate).
}
\description{
This function allows to use a custom thresholding method to minimize the Miss-Rate (False Negative Rate). A data.table of values is returned.
}

