% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_precision.R
\name{plotting.max_precision}
\alias{plotting.max_precision}
\title{Maximum Precision (Positive Predictive Value) plotting}
\usage{
plotting.max_precision(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their Precision (Positive Predictive Value).
}
\description{
This function allows to use a custom thresholding method to maximize the Precision (Positive Predictive Value). A data.table of values is returned.
}

