% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.max_specificity.R
\name{plotting.max_specificity}
\alias{plotting.max_specificity}
\title{Maximum Specificity (True Negative Rate) plotting}
\usage{
plotting.max_specificity(preds, labels, plots = TRUE, ...)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{plots}{Type: numeric. Whether to plot the data immediately or not.}

\item{...}{Other arguments to pass to \code{plot}.}
}
\value{
A data.table containing the probabilities and their Specificity (True Negative Rate).
}
\description{
This function allows to use a custom thresholding method to maximize the Specificity (True Negative Rate). A data.table of values is returned.
}

