% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LauraeML_utils.R
\name{print_multi}
\alias{print_multi}
\title{Print appropriately formatted hyperparameters and error}
\usage{
print_multi(params)
}
\arguments{
\item{params}{Type: list of vectors. A list of vectors formatted like \code{list(c("alpha", 1.248, 2, 5), c("beta", 2.58, 2, 5), c("integer", 8, 2, -1))}, which is here (alpha = 1.248 to print with 2 decimals and 5 digits => "01.24800") (beta = 2.58 to print with 2 digits and 5 decimals => "02.58000") (integer = 8 to print with 2 decimals => "08").}
}
\value{
The formatted double precision value.
}
\description{
This function is a helper function to print the hyperparameters and error appropriately in a simple call. The first element refers to the name, the second is for the value, the third for the number of digits, and the fourth for the number of decimals. Specify the last parameter as \code{-1} in order to pass the value as integer.
}
\examples{
best_params <- list(c("alpha", 1.248, 2, 5), c("beta", 2.58, 2, 5), c("integer", 8, 2, -1))
print_multi(best_params) # "alpha=01.24800, beta=02.58000, integer=08"

}

