% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.max_mcc.R
\name{prob.max_mcc}
\alias{prob.max_mcc}
\title{Probability Matthews Correlation Coefficient}
\usage{
prob.max_mcc(preds, labels, thresh = 0.5)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{thresh}{Type: numeric. The cutoff (threshold) probability which is deemed positive when higher or equal, or negative when lower. Defaults to \code{0.5} (anything higher or equal to 0.5 is positive, anything lower is negative).}
}
\value{
The Matthews Correlation Coefficient at the provided threshold.
}
\description{
This function allows to use a custom thresholding method to compute the Matthews Correlation Coefficient.
}

