% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse_csv.R
\name{read_sparse_csv}
\alias{read_sparse_csv}
\title{Read sparse (numeric) CSVs}
\usage{
read_sparse_csv(input, iterfeature, nfeatures = NA, colClasses = NA,
  RDS = NA, compress_RDS = TRUE, NA_sparse = FALSE)
}
\arguments{
\item{input}{The input file name.}

\item{iterfeature}{The amount of variables loaded per iteration. The smaller the longer it takes to load the whole dataset in its entireity.}

\item{nfeatures}{The IDs of features to load. Defaults to \code{NA} which means loading all columns.}

\item{colClasses}{The classes of the columns. Defaults to \code{NA} which means autoselection as numeric. Do not modify (keep default).}

\item{RDS}{Whether to store in a RDS file of that name. Defaults to \code{NA} which means no RDS file. Otherwise, it takes \code{RDS} as filename.}

\item{compress_RDS}{Whether to compress RDS file. Defaults to \code{TRUE}}

\item{NA_sparse}{Whether sparsity is defined as NA. Defaults to \code{FALSE}}
}
\value{
The sparse matrix
}
\description{
This function allows you to big load sparse numeric CSVs. Loading in chunks allows to not explode the memory as when the data is imported into R, it is typically a dense matrix.
Verbosity is automatic and cannot be removed. In case you need this function without verbosity, please compile the package after removing verbose messages.
}
\examples{
read_sparse_csv("train_numeric.csv", iterfeature = 100, IDs = c(1:500, 601:1000), colClasses = NA,
RDS = TRUE, compress_RDS = FALSE, NA_sparse = FALSE)

}

