% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.lm.R
\name{report.lm}
\alias{report.lm}
\title{Linear Regression Modeling HTML report}
\usage{
report.lm(data, label, folds, normalize = TRUE, cleaning = TRUE,
  deficiency = TRUE, stats = TRUE, coefficients = TRUE,
  adv_stats = TRUE, plots = TRUE, output_file = "report.lm.html",
  output_dir = getwd(), open_file = TRUE, ...)
}
\arguments{
\item{data}{Type: data.table. The data to fit a linear regression model on.}

\item{label}{Type: vector. The label the data must fit to.}

\item{folds}{Type: list of numeric vectors. The folds used.}

\item{normalize}{Type: boolean. Whether features should be normalized before being fed to the linear model. Defaults to \code{TRUE}.}

\item{cleaning}{Type: boolean. Whether NAs are set to 0 (if there are any). Defaults to \code{TRUE}.}

\item{deficiency}{Type: boolean. Whether rank deficiency computation is done (kappa). Defaults to \code{TRUE}.}

\item{stats}{Type: boolean. Whether machine learning statistics should be output for model performance diagnosis. When \code{TRUE}, also returns the metrics and the out of fold predictions. Defaults to \code{TRUE}.}

\item{coefficients}{Type: boolean. Whether feature coefficients should be output. Defaults to \code{TRUE}.}

\item{adv_stats}{Type: boolean. Whether advanced statistics should be done to analyze in depth the data. Defaults to \code{TRUE}.}

\item{plots}{Type: boolean. Whether plotting of fitted values vs predicted values should be done. Defaults to \code{TRUE}.}

\item{output_file}{Type: character. The output report file name. Defaults to \code{"report.lm.html"}.}

\item{output_dir}{Type: character. The output report directory name. Defaults to \code{getwd()}.}

\item{open_file}{Type: boolean. Whether to open the output report once it has finished computing. Defaults to \code{TRUE}.}

\item{...}{Other arguments to pass to \code{rmarkdown::render}.}
}
\value{
Returns a list with the machine learning metrics (\code{"Metrics"}), the folds \code{"Folds"}, the fitted values per fold (\code{"Fitted"}), the predicted values per fold (\code{"Predicted"}) if they were computed. Otherwise, returns \code{TRUE}.
}
\description{
This function creates a linear regression report as a HTML file. Cross-validation is possible.
}
\examples{
# No example.
\dontrun{
  library(rmarkdown)
  library(DT)
  library(formattable)
  library(RcppArmadillo)
}

}

