% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.xgb.helper.R
\name{report.xgb.helper}
\alias{report.xgb.helper}
\title{Extreme Gradient Boosting HTML report helper function}
\usage{
report.xgb.helper(train, valid, params, nrounds = 1e+06, model = NULL)
}
\arguments{
\item{train}{Type: xgb.DMatrix. The labelled training data.}

\item{valid}{Type: xgb.DMatrix. The labelled testing data.}

\item{params}{Type: list. The parameters as a list.}

\item{nrounds}{Type: numeric. The number of training iterations.}

\item{model}{Type: xgb.Booster. The xgboost model to train from.}
}
\value{
A xgboost model.
}
\description{
This function is used to train a model using xgboost in the automated HTML report creator. It must handle three arguments: \code{train}, \code{valid}, and \code{params}.
}
\examples{
# No example.
\dontrun{
}

}

