% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.xgb.R
\name{report.xgb}
\alias{report.xgb}
\title{Extreme Gradient Boosting HTML report}
\usage{
report.xgb(data, label, folds, params, normalize = TRUE,
  classification = TRUE, importance = TRUE, unbiased = TRUE,
  stats = TRUE, plots = TRUE, output_file = "report.xgb.html",
  output_dir = getwd(), open_file = TRUE, ...)
}
\arguments{
\item{data}{Type: data.table. The data to fit a xgboost model on.}

\item{label}{Type: vector. The label the data must fit to.}

\item{folds}{Type: list of numeric vectors. The folds used.}

\item{params}{Type: list. The parameters to pass to \code{report.xgb.helper}.}

\item{normalize}{Type: boolean. Whether features should be normalized before being fed to the xgboost model. Defaults to \code{TRUE}.}

\item{classification}{Type: boolean. Whether the task is a classification or not. Defaults to \code{TRUE}.}

\item{importance}{Type: boolean. Whether to perform feature importance computation or not. Defaults to \code{TRUE}.}

\item{unbiased}{Type: boolean. Whether to perform unbiased feature importance computation or not. Defaults to \code{TRUE}.}

\item{stats}{Type: boolean. Whether machine learning statistics should be output for model performance diagnosis. When \code{TRUE}, also returns the metrics and the out of fold predictions. Defaults to \code{TRUE}.}

\item{plots}{Type: boolean. Whether plotting of fitted values vs predicted values should be done. Defaults to \code{TRUE}.}

\item{output_file}{Type: character. The output report file name. Defaults to \code{"report.lm.html"}.}

\item{output_dir}{Type: character. The output report directory name. Defaults to \code{getwd()}.}

\item{open_file}{Type: boolean. Whether to open the output report once it has finished computing. Defaults to \code{TRUE}.}

\item{...}{Other arguments to pass to \code{rmarkdown::render}.}
}
\value{
Returns a list with the machine learning metrics (\code{"Metrics"}), the machine learning probabilities (\code{"Probs"}), the folds \code{"Folds"}, the fitted values per fold (\code{"Fitted"}), the predicted values per fold (\code{"Predicted"}), the biased feature importance (\code{"BiasedImp"}), and the unbiased feature importance (\code{"UnbiasedImp"}) if they were computed. Otherwise, returns \code{TRUE}.
}
\description{
This function creates an xgboost as a HTML file. Cross-validation is mandatory. Does NOT handle multiclass scenarios or non-regression/classification tasks. Does NOT handle gblinear. You cannot use \code{process_type}, \code{updater}, and \code{refresh_leaf} parameters.
}
\examples{
# No example.
\dontrun{
  library(Laurae)
  library(data.table)
  library(rmarkdown)
  library(xgboost)
  library(DT)
  library(formattable)
  library(matrixStats)
  library(lattice)
  library(R.utils)
}

}

