% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_single.R
\name{rule_single}
\alias{rule_single}
\title{Outlying univariate continuous association rule finder}
\usage{
rule_single(data, label, train_rows = length(label), iterations = 1000,
  minimal_score = 25, minimal_node = 5, false_negatives = 2,
  seed = 11111, scoring = TRUE, ruling = TRUE)
}
\arguments{
\item{data}{The data.frame containing the features to make association rules on, or the scoring matrix. Missing values are not allowed.}

\item{label}{The target label as an integer vector (each value must be either 0 or 1). 1 must be the miniority label.}

\item{train_rows}{The rows used for training the association rules. Must be your training set, whose length is equal to \code{length(labels)}. Defaults to \code{length(label)}.}

\item{iterations}{The amount of iterations allowed for limited-memory Gradient Descent}

\item{minimal_score}{The association rule finder will not accept any node under the allowed outlying score. Defaults to \code{25}.}

\item{minimal_node}{The association rule finder will not accept any node containing under that specific amount of samples. Defaults to \code{5}.}

\item{false_negatives}{The association rule will allow at most (\code{false_negatives - 1}) false negatives. A higher allows a more permissive algorithm, lower makes it very difficult to converge (or to find any rule at all). Defaults to \code{2}.}

\item{seed}{The random seed for reproducibility. Defaults to \code{11111}.}

\item{scoring}{Whether to score features before computing the association rules. Defaults to \code{TRUE}.}

\item{ruling}{Whether to rule features (useful when you only want the scores). Defaults to \code{TRUE}.}
}
\value{
A list with one to three elements: \code{"scores"} the outlying scores for features, \code{"parsed_scores"} for the association rule result on specific features, and \code{"output"} for the association rule general result per observation.
}
\description{
This function allows you to search for association rules on outlying univariate continuous features against a binary label. The predicted label is 0, and the overfitting severity is very high (see: Kaggle's Santander Customer Satisfaction competition).
It can be used to score outliers first, then make rules afterwards if needed.
Verbosity is automatic and cannot be removed. In case you need this function without verbosity, please compile the package after removing verbose messages.
}
\examples{
\dontrun{
scored_data <- rule_single(data = data, label = NA, scoring = TRUE, ruling = FALSE)
rules <- rule_single(data = scored_data, label = target,
iterations = 100, scoring = FALSE, ruling = TRUE)
preds <- preds[rules$output[(length(target)+1):(nrow(data))] == 0] <- 0
}

}

