% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDF.R
\name{setDF}
\alias{setDF}
\title{Convert data.table to data.frame without copy}
\usage{
setDF(x)
}
\arguments{
\item{x}{Type: data.table. The data.table to convert to data.frame.}
}
\value{
Nothing (the data.table is transformed to a data.frame already)
}
\description{
This function allows to coerce a data.table to a data.frame without making in-memory copy. Therefore, you do not need 3x your current memory to coerce to a data.frame from a data.table.
}
\examples{
#library(data.table)
df <- data.frame(matrix(nrow = 100, ncol = 10))
coln <- colnames(df)
rown <- rownames(df)
setDT(df)
is.data.table(df) # Returns TRUE = OK
sum(c(coln == colnames(df), c(rown == rownames(df)))) # Returns 2 = OK
setDF(df)
is.data.table(df) # Returns FALSE = OK
sum(c(coln == colnames(df), c(rown == rownames(df)))) # Returns 2 = OK

}

