% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_smooth_func.R
\name{stat_smooth_func}
\alias{stat_smooth_func}
\title{ggplot facet function with printed formula (non Plotly)}
\usage{
stat_smooth_func(mapping = NULL, data = NULL, geom = "smooth",
  position = "identity", ..., method = "auto", formula = y ~ x,
  se = TRUE, n = 80, span = 0.75, fullrange = FALSE, level = 0.95,
  method.args = list(), na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, xpos = NULL, ypos = NULL)
}
\arguments{
\item{mapping}{...}

\item{data}{...}

\item{geom}{...}

\item{position}{...}

\item{...}{...}

\item{method}{...}

\item{formula}{...}

\item{se}{...}

\item{n}{...}

\item{span}{...}

\item{fullrange}{...}

\item{level}{...}

\item{method.args}{...}

\item{na.rm}{...}

\item{show.legend}{...}

\item{inherit.aes}{...}

\item{xpos}{...}

\item{ypos}{...}
}
\value{
See stat_smooth.
}
\description{
This function prints linear formula used on each facet for ggplot. Refer to \code{stat_smooth} for description. Does not support Plotly.
}
\details{
See this gist: https://gist.github.com/kdauria/524eade46135f6348140

Extra edits were made to fit the needs of this package.
}

