% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableplot_jpg.R
\name{tableplot_jpg}
\alias{tableplot_jpg}
\title{Batch tableplot generator to JPEG}
\usage{
tableplot_jpg(data, target, nBins = 100, folder = "./autoplots/plot_",
  ID = FALSE, width = 960, height = 960, pointsize = 12)
}
\arguments{
\item{data}{The data to load.}

\item{target}{The target label. Either factor or numeric.}

\item{nBins}{The amount of bins per histogram. Should be between 50 and 500 if possible (higher may overfit, lower may underfit). Defaults to \code{100}.}

\item{folder}{The output folder where tableplots will be stored. Defaults to \code{"./autoplots/plot_"}.}

\item{ID}{Should the filename use the ID of the feature (starts at 1) or the name of the feature? Defaults to \code{FALSE}.}

\item{width}{The width output of each tableplot, in pixels. Defaults to \code{960}.}

\item{height}{The height output of each tableplot, in pixels. Defaults to \code{960}.}

\item{pointsize}{The pointsize output of each tableplot, in pixels. Defaults to \code{12}.}
}
\value{
Nothing
}
\description{
This functions allows you to create JPEGs of tableplots aganist a dataset and a specific label (which must be a factor or a numeric)
A numeric target allows to provide an overview of each variable against the ordering of the label
A factor target allows to provide an overview of each variable against each factor of the label
}
\examples{
#tableplot_jpg(train[, !(colnames(train) \%in\% "target")], train$target,
#folder = "./plots/my_plot_", ID = FALSE, width = 960, height = 960, pointsize = 12)

}

