% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{Get current Time in Milliseconds}
\usage{
timer()
}
\value{
The current time at the call in milliseconds.
}
\description{
This function returns the current time in milliseconds.
}
\examples{
library(R.utils)
timer()

}

