% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer_func_print.R
\name{timer_func_print}
\alias{timer_func_print}
\title{Get Function Time in Milliseconds (with printing)}
\usage{
timer_func_print(expr, seconds = FALSE, pre_msg = "The function ran in ",
  msg_format = "\%05.03f", post_msg = ifelse(seconds == TRUE, " seconds.",
  " milliseconds."), linebreak = TRUE)
}
\arguments{
\item{expr}{Type: any expression. An expression to evalutate (the expression you want to benchmark).}

\item{seconds}{Type: logical. Whether you want milliseconds (default) or seconds. Defaults to \code{FALSE}.}

\item{pre_msg}{Type: character. The message preceeding the time print. Defaults to \code{"The function ran in "}.}

\item{msg_format}{Type: character. The formatting applied to the evaluation time, used by \code{sprintf}. Defaults to \code{"\%05.03f"}.}

\item{post_msg}{Type: character. The message succeeding the time print. Defaults to \code{ifelse(seconds == TRUE, " seconds.", " milliseconds."}.}

\item{linebreak}{Type: logical. Whether to add a line break after the printed message. Defaults to \code{TRUE}.}
}
\value{
The time needed for a function to run in milliseconds (or seconds).
}
\description{
This function returns the time needed for a function to run in milliseconds, and prints it.
}
\examples{
library(R.utils)
timer_func_print({for (i in 1:100) {rnorm(1)}}, seconds = TRUE)

}

