% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_grid.R
\name{tsne_grid}
\alias{tsne_grid}
\title{t-SNE grid search function}
\usage{
tsne_grid(data, output_dims, input_dims = ncol(data),
  perplexity_range = c(1, floor((ncol(data) - 1)/3)), tries = 10,
  iterations = 10000, theta = 0, check_duplicates = FALSE, pca = FALSE,
  is_distance = FALSE)
}
\arguments{
\item{data}{The data.frame input into t-SNE}

\item{output_dims}{How many dimensions to output? (increases exponentially the computation time)}

\item{input_dims}{How many input dimensions to use? (defaults to \code{ncol(data)}) - this should be changed when using pca to a value below the default value}

\item{perplexity_range}{What hyperparameter interval to look for? (should be formatted as (min, max)) - defaults to \code{c(1, floor((ncol(data)-1)/3))} - to grid search a seed for a fixed perplexity value, use min = max as inputs - the best pragmatic perpelxity for the lowest loss is typically \code{floor((ncol(data)-1)/3)}}

\item{tries}{How many seeds to test t-SNE per perplexity value? (this increases linearly the computation time)}

\item{iterations}{How many iterations per t-SNE are performed? (this increases approximately linearly the computation time)}

\item{theta}{Use exact t-SNE (0) or Barnes-Hut t-SNE? (in ]0, 1] interval)}

\item{check_duplicates}{Should t-SNE check for duplicates? (unlike common beliefs, t-SNE works perfectly with the existance of identical observations)}

\item{pca}{Should a PCA (Principal Component Analysis) be performed? (note: it is performed every iteration, therefore it is computationally intensive and should be avoided - if you need PCA, please input the PCA instead of the data)}

\item{is_distance}{Is the input a distance matrix? (assumes the diagonal cuts in half the input data.frame)}
}
\value{
A list with the best (lowest loss at a specific iteration) t-SNE elements from Rtsne
}
\description{
This function allows you to search a perplexity hyperparameter range along with different seeds.
Verbosity is automatic and cannot be removed. In case you need this function without verbosity, please compile the package after removing verbose messages.
}
\examples{
#tsne_model <- tsne_grid(initial_diag = initial_diag, dims = 3,
#perplexity_range = c(floor((ncol(initial_diag)-1)/3), floor((ncol(initial_diag)-1)/3)),
#tries = 100, iterations = 10000, theta = 0.0, check_duplicates = FALSE,
#pca = FALSE, is_distance = TRUE)

}

