% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.max_f1.R
\name{xgb.max_f1}
\alias{xgb.max_f1}
\title{xgboost evaluation metric for maximum F1 Score (Precision with Sensitivity harmonic mean)}
\usage{
xgb.max_f1(pred, dtrain)
}
\arguments{
\item{pred}{Type: numeric. The predictions.}

\item{dtrain}{Type: xgb.DMatrix. The training data.}
}
\value{
The maximum F1 Score (Precision with Sensitivity harmonic mean) for binary data.
}
\description{
This function allows xgboost to use a custom thresholding method to maximize the F1 Score (Precision with Sensitivity harmonic mean). You can use this function via \code{eval_metric}. It leaks memory over time, but it can be reclaimed using \code{gc()}.
}

