% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.max_fallout.R
\name{xgb.max_fallout}
\alias{xgb.max_fallout}
\title{xgboost evaluation metric for minimum Fall-Out (False Positive Rate)}
\usage{
xgb.max_fallout(pred, dtrain)
}
\arguments{
\item{pred}{Type: numeric. The predictions.}

\item{dtrain}{Type: xgb.DMatrix. The training data.}
}
\value{
The minimum Fall-Out (False Positive Rate) for binary data.
}
\description{
This function allows xgboost to use a custom thresholding method to minimize the Fall-Out (False Positive Rate). You can use this function via \code{eval_metric}. It leaks memory over time, but it can be reclaimed using \code{gc()}.
}

