% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.max_missrate.R
\name{xgb.max_missrate}
\alias{xgb.max_missrate}
\title{xgboost evaluation metric for minimum Miss-Rate (False Negative Rate)}
\usage{
xgb.max_missrate(pred, dtrain)
}
\arguments{
\item{pred}{Type: numeric. The predictions.}

\item{dtrain}{Type: xgb.DMatrix. The training data.}
}
\value{
The minimum Miss-Rate (False Negative Rate) for binary data.
}
\description{
This function allows xgboost to use a custom thresholding method to minimum the Miss-Rate (False Negative Rate). You can use this function via \code{eval_metric}. It leaks memory over time, but it can be reclaimed using \code{gc()}.
}

