% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.max_precision.R
\name{xgb.max_precision}
\alias{xgb.max_precision}
\title{xgboost evaluation metric for maximum Precision (Positive Predictive Value)}
\usage{
xgb.max_precision(pred, dtrain)
}
\arguments{
\item{pred}{Type: numeric. The predictions.}

\item{dtrain}{Type: xgb.DMatrix. The training data.}
}
\value{
The maximum Precision (Positive Predictive Value) for binary data.
}
\description{
This function allows xgboost to use a custom thresholding method to maximize the Precision (Positive Predictive Value). You can use this function via \code{eval_metric}. It leaks memory over time, but it can be reclaimed using \code{gc()}.
}

