% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.max_specificity.R
\name{xgb.max_specificity}
\alias{xgb.max_specificity}
\title{xgboost evaluation metric for maximum Specificity (True Negative Rate)}
\usage{
xgb.max_specificity(pred, dtrain)
}
\arguments{
\item{pred}{Type: numeric. The predictions.}

\item{dtrain}{Type: xgb.DMatrix. The training data.}
}
\value{
The maximum Specificity (True Negative Rate) for binary data.
}
\description{
This function allows xgboost to use a custom thresholding method to maximize the Specificity (True Negative Rate) You can use this function via \code{eval_metric}. It leaks memory over time, but it can be reclaimed using \code{gc()}.
}

