% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboard.init.R
\name{xgboard.init}
\alias{xgboard.init}
\title{Xgboard Metric Evaluation Initialization (Environment)}
\usage{
xgboard.init(what, watchnames, maximizer, log)
}
\arguments{
\item{what}{Type: character. A character or a vector of characters corresponding to the names of the metrics to log (in the right order, if there are multiples). It must be the same length of \code{eval_metric} passed to \code{param} on \code{xgb.train}. Do not use more than 3 metrics, as Xgboard is designed for a maximum of 3 metrics.}

\item{watchnames}{Type: character. A character or a vector of characters corresponding to the names of the watchlist which are evaluated against the metric to log (in the right order, if there are multiples). It must be the same length of \code{watchlist} passed to \code{xgb.train}.}

\item{maximizer}{Type: logical. A logical or a vector of logicals corresponding to whether to maximize (\code{TRUE}) or minimize (\code{FALSE}) the evaluation metrics (\code{what}). It must be the same length of \code{what}.}

\item{log}{Type: character. Where to store the dump/log file? It must be an absolute path.}
}
\value{
An environment for the dumping. Does not store the dump though.
}
\description{
This function initializes the environment for the dump (logging). This must be run before any xgboost run.
}
\examples{
\dontrun{
# We prepare environment for Accuracy/Threshold logging on Train/Test
# Stored in D:/debug/log.txt
my_envir <- xgboard.init(what = c("Accuracy", "Threshold"),
                         watchnames = c("Train", "Test"),
                         maximizer = c(TRUE, TRUE),
                         log = "D:/debug/log.txt")
}

}

