% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboard.xgb.R
\name{xgboard.xgb}
\alias{xgboard.xgb}
\title{Xgboard Metric Evaluation Creator (Wrapper)}
\usage{
xgboard.xgb(..., f, dumper)
}
\arguments{
\item{...}{Type: nothing. Ignore.}

\item{f}{Type: function. An xgboost evaluation metric function with \code{dump} input.}

\item{dumper}{Type: environment. An environment created by \code{xgboard.init}.}
}
\description{
This function is a wrapper to create an evaluation metric.
}
\examples{
\dontrun{
# Model will be trained using these parameters
# Take note of eval_metric needing xgboard.xgb(f = your metric, dumper = envir)
param <- list(max_depth = 2,
              eta = 1,
              silent = 1,
              nthread = 2, 
              objective = "binary:logistic",
              eval_metric = xgboard.xgb(f = board.evalerror, dumper = my_envir))
}

}

