% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTcolsample.R
\name{DTcolsample}
\alias{DTcolsample}
\title{data.table colsampling (nearly without) copy}
\usage{
DTcolsample(DT, kept, remove = FALSE)
}
\arguments{
\item{DT}{Type: data.table. The data.table to combine on.}

\item{kept}{Type: vector of integers or vector of characters. The columns to select to keep.}

\item{remove}{Type: boolean. Whether the argument \code{kept} acts as a removal (keep all columns which are not in \code{kept}). Defaults to \code{FALSE}.}
}
\value{
The subsampled data.table.
}
\description{
This function attempts to subsample one data.table without making copies. Well, you could just use \code{DT[, (mycols) := NULL]} (for removal) or \code{DT <- DT[, (mycols), with = FALSE]} for selecting...
}
\details{
Warning: \code{DT} is a pointer only even if you pass the object to this function. This is how memory efficiency is achieved.
}
\examples{
library(data.table)
DT <- data.frame(matrix(nrow = 50, ncol = 10))
DT <- setDT(DT)
colnames(DT) <- paste(colnames(DT), "xx", sep = "")
DT <- DTcolsample(DT, kept = 1:8, remove = FALSE)
DT <- DTcolsample(DT, kept = 1:6, remove = TRUE)

}

