% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbm.metric.R
\name{lgbm.metric}
\alias{lgbm.metric}
\title{LightGBM Metric Output}
\usage{
lgbm.metric(workingdir, log_name, metrics = TRUE, data.table = TRUE)
}
\arguments{
\item{workingdir}{Type: character. The working directory of the model file. Defaults to \code{ifelse(is.list(model), model[["Path"]], getwd())}, which means "take the model working directory if provided the model list, else take the default working directory".}

\item{log_name}{Type: character. The logging (sink) file to output (like 'log.txt'). Defaults to \code{NA}.}

\item{metrics}{Type: boolean. Whether to return the metrics table (\code{TRUE}) or the best iteration (\code{FALSE}).}

\item{data.table}{Type: boolean. Whether to return a data.table (\code{TRUE}) or a data.frame (\code{FALSE}). Defaults to \code{TRUE}. Useless when \code{metrics} is set to \code{FALSE}.}
}
\value{
A data.table (or data.frame) with 3 to 5 columns: \code{c("Iteration", "Total_Time", "Round_Time", "Train_Loss", "Test_Loss")}.
}
\description{
This function allows to get the metric values from a LightGBM log.
}
\examples{
\dontrun{
# Get the metrics of the model trained (requires verbose = FALSE during training)
lgbm.metric(workingdir = trained[["Path"]],
log_name = trained[["Log"]],
metrics = TRUE,
data.table = TRUE)

# Get the best iteration of the model trained (requires verbose = FALSE during training)
lgbm.metric(workingdir = trained[["Path"]],
log_name = trained[["Log"]],
metrics = FALSE)
}

}

