% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MAE_grad}
\alias{loss_MAE_grad}
\title{Mean Absolute Error (gradient function)}
\usage{
loss_MAE_grad(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The gradient of the Absolute Error per value.
}
\description{
This function computes the Mean Absolute Error loss (MAE) gradient per value provided \code{preds} and \code{labels}.
}

