% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MAE_math}
\alias{loss_MAE_math}
\title{Mean Absolute Error (symbolic function)}
\usage{
loss_MAE_math(x, y)
}
\arguments{
\item{x}{The \code{predictions}.}

\item{y}{The \code{labels}.}
}
\value{
The Squared Error per value.
}
\description{
This function computes the Mean Absolute Error loss (MAE) per value provided \code{x} (preds - labels).
}
\details{
Supposing: \eqn{x = preds - labels}

Loss Formula : \eqn{abs(x)}

Gradient Formula : \eqn{sign(x)}

Hessian Formula : \eqn{0}
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_MAE_math)
}

}

