% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MAE_xgb}
\alias{loss_MAE_xgb}
\title{Mean Absolute Error (xgboost function)}
\usage{
loss_MAE_xgb(preds, dtrain)
}
\arguments{
\item{preds}{The predictions.}

\item{dtrain}{The labels.}
}
\value{
The gradient and the hessian of the Absolute Error per value in a list.
}
\description{
This function computes for xgboost's \code{obj} function the Mean Absolute Error loss (MAE) gradient and hessian per value provided \code{preds} and \code{dtrain}.
}

