% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MCE_grad}
\alias{loss_MCE_grad}
\title{Mean Cubic Error (gradient function)}
\usage{
loss_MCE_grad(y_pred, y_true)
}
\arguments{
\item{y_pred}{The \code{predictions}.}

\item{y_true}{The \code{labels}.}
}
\value{
The gradient of the Cubic Error per value.
}
\description{
This function computes the Mean Cubic Error loss (MCE) gradient per value provided \code{preds} and \code{labels}.
}
\details{
Supposing: \eqn{x = preds - labels}

Loss Formula : \eqn{abs(x^3)}

Gradient Formula : \eqn{3 * (x * abs(x))}

Hessian Formula : \eqn{(6 * x * x) / abs(x)}
}

