% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MCE_hess}
\alias{loss_MCE_hess}
\title{Mean Cubic Error (hessian function)}
\usage{
loss_MCE_hess(y_pred, y_true)
}
\arguments{
\item{y_pred}{The predictions.}

\item{y_true}{The labels.}
}
\value{
The hessian of the Cubic Error per value.
}
\description{
This function computes the Mean Cubic Error loss (MCE) hessian per value provided \code{preds} and \code{labels}.
}

