% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MCE_math}
\alias{loss_MCE_math}
\title{Mean Cubic Error (math function)}
\usage{
loss_MCE_math(x)
}
\arguments{
\item{x}{The difference between the \code{label} and the \code{prediction}.}
}
\value{
The Cubic Error per value.
}
\description{
This function computes the Mean Cubic Error loss (MCE) per value provided \code{x} (preds - labels).
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_MCE_math, fc_ref)
}

}

