% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MSE_math}
\alias{loss_MSE_math}
\title{Mean Squared Error (symbolic function)}
\usage{
loss_MSE_math(x)
}
\arguments{
\item{x}{The difference between the \code{label} and the \code{prediction}.}
}
\value{
The Squared Error per value.
}
\description{
This function computes the Mean Squared Error loss (MSE) per value provided \code{x} (preds - labels).
}
\examples{
\dontrun{
SymbolicLoss(fc = loss_MSE_math, fc_ref)
}

}

