% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossFunctions.R
\name{loss_MSE_xgb}
\alias{loss_MSE_xgb}
\title{Mean Squared Error (xgboost function)}
\usage{
loss_MSE_xgb(preds, dtrain)
}
\arguments{
\item{preds}{The \code{predictions}.}

\item{dtrain}{The xgboost model.}
}
\value{
The gradient and the hessian of the Squared Error per value in a list.
}
\description{
This function computes for xgboost's \code{obj} function the Mean Squared Error loss (MSE) gradient and hessian per value provided \code{preds} and \code{dtrain}.
}
\details{
Supposing: \eqn{x = preds - labels}

Loss Formula : \eqn{x^2}

Gradient Formula : \eqn{2 * x}

Hessian Formula : \eqn{2}
}

